requirements.txt
google-genai==1.38.0
pillow==11.3.0
pydantic==2.11.9
rich==14.1.0